// IsveskiDemo_NodeJS/clientcode/model/baseAction.ts
/**
 * Isveski Client
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ActionTypesEnum } from './actionTypesEnum';

export class BaseAction {
    'actionType': ActionTypesEnum;

    static discriminator: string | undefined = "actionType";

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "actionType",
            "baseName": "actionType",
            "type": "ActionTypesEnum"
        }    ];

    static getAttributeTypeMap() {
        return BaseAction.attributeTypeMap;
    }
}

export namespace BaseAction {
}