// IsveskiDemo_NodeJS/clientcode/model/clientTicketAction.ts
/**
 * Isveski Client
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ActionType } from './actionType';

export class ClientTicketAction {
    'id'?: string;
    'actionType'?: ActionType;
    'name'?: string | null;
    'data'?: string | null;
    'presentationData'?: string | null;
    'presentationType'?: string | null;
    'actionTime'?: Date;
    'note'?: string | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "actionType",
            "baseName": "actionType",
            "type": "ActionType"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "data",
            "baseName": "data",
            "type": "string"
        },
        {
            "name": "presentationData",
            "baseName": "presentationData",
            "type": "string"
        },
        {
            "name": "presentationType",
            "baseName": "presentationType",
            "type": "string"
        },
        {
            "name": "actionTime",
            "baseName": "actionTime",
            "type": "Date"
        },
        {
            "name": "note",
            "baseName": "note",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return ClientTicketAction.attributeTypeMap;
    }
}

export namespace ClientTicketAction {
}